+++ web/app/documentation/create_text_index.js
@@ -1,3 +1,0 @@
-db.themes.createIndex({ "name": "text", "description": "text" })
-
-db.slates.createIndex({ "options.searchName": "text", "options.searchDescription": "text", "options.searchText": "text" });